+++ mrserver/main_page.go
@@ -1,8 +1,0 @@
-package mrserver
-
-import "net/http"
-
-func MainPage(w http.ResponseWriter, r *http.Request) {
-	w.WriteHeader(http.StatusOK)
-	w.Write([]byte("{\"STATUS\": \"OK\"}"))
-}