+++ mrtype/image.go
@@ -1,24 +1,20 @@
 package mrtype
 
 import (
-	"io"
 	"time"
 )
 
 type (
-	FileInfo struct {
+	ImageInfo struct {
 		ContentType  string     `json:"contentType,omitempty"`
 		OriginalName string     `json:"originalName,omitempty"`
 		Name         string     `json:"name,omitempty"`
 		Path         string     `json:"-"`
 		URL          string     `json:"url,omitempty"`
+		Width        int32      `json:"width,omitempty"`
+		Height       int32      `json:"height,omitempty"`
 		Size         int64      `json:"size,omitempty"`
 		CreatedAt    *time.Time `json:"created,omitempty"`
 		ModifiedAt   *time.Time `json:"lastModified,omitempty"`
-	}
-
-	File struct {
-		FileInfo
-		Body io.ReadCloser
 	}
 )