+++ mrtype/string_pointer.go
@@ -1,0 +1,9 @@
+package mrtype
+
+func StringPointer(value string) *string {
+	if value == "" {
+		return nil
+	}
+
+	return &value
+}