+++ mrtype/time_pointer.go
@@ -1,0 +1,11 @@
+package mrtype
+
+import "time"
+
+func TimePointer(value time.Time) *time.Time {
+	if value.IsZero() {
+		return nil
+	}
+
+	return &value
+}